split out the functionality of the RestApplication class
